requirements.txt
python-telegram-bot
skodaconnect